docs and test
