+++ mkdocs.yml
@@ -123,6 +123,7 @@ nav:
     - Multiple output cache: multiple_output_cache.md
     - Sandbox: sandbox.md
     - Generating recipes: recipe_generation.md
+    - Creating patches: create_patch.md
     - Tips and tricks: tips_and_tricks.md
 
   - Testing:
